remove code from model/__init__.py
