+++ backend/tracim_backend/fixtures/users_and_groups.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
-from tracim_backend import models
+from tracim_backend.models.auth import User
+from tracim_backend.models.auth import Group
 from tracim_backend.fixtures import Fixture
 from tracim_backend.lib.core.user import UserApi
 
@@ -8,7 +9,7 @@ class Base(Fixture):
     require = []
 
     def insert(self):
-        u = models.User()
+        u = User()
         u.display_name = 'Global manager'
         u.email = 'admin@admin.admin'
         u.password = 'admin@admin.admin'
@@ -19,21 +20,21 @@ class Base(Fixture):
             current_user=u)
         uapi.execute_created_user_actions(u)
 
-        g1 = models.Group()
+        g1 = Group()
         g1.group_id = 1
         g1.group_name = 'users'
         g1.display_name = 'Users'
         g1.users.append(u)
         self._session.add(g1)
 
-        g2 = models.Group()
+        g2 = Group()
         g2.group_id = 2
         g2.group_name = 'trusted-users'
         g2.display_name = 'Trusted Users'
         g2.users.append(u)
         self._session.add(g2)
 
-        g3 = models.Group()
+        g3 = Group()
         g3.group_id = 3
         g3.group_name = 'administrators'
         g3.display_name = 'Administrators'
@@ -45,26 +46,26 @@ class Test(Fixture):
     require = [Base, ]
 
     def insert(self):
-        g2 = self._session.query(models.Group).\
-            filter(models.Group.group_name == 'trusted-users').one()
+        g2 = self._session.query(Group).\
+            filter(Group.group_name == 'trusted-users').one()
 
-        lawrence = models.User()
+        lawrence = User()
         lawrence.display_name = 'Lawrence L.'
         lawrence.email = 'lawrence-not-real-email@fsf.local'
         lawrence.password = 'foobarbaz'
         self._session.add(lawrence)
         g2.users.append(lawrence)
 
-        bob = models.User()
+        bob = User()
         bob.display_name = 'Bob i.'
         bob.email = 'bob@fsf.local'
         bob.password = 'foobarbaz'
         self._session.add(bob)
         g2.users.append(bob)
 
-        g1 = self._session.query(models.Group).\
-            filter(models.Group.group_name == 'users').one()
-        reader = models.User()
+        g1 = self._session.query(Group).\
+            filter(Group.group_name == 'users').one()
+        reader = User()
         reader.display_name = 'John Reader'
         reader.email = 'john-the-reader@reader.local'
         reader.password = 'read'
