+++ backend/tracim_backend/models/context_models.py
@@ -20,7 +20,7 @@ from tracim_backend.lib.utils.utils import string_to_list
 from tracim_backend.lib.utils.utils import WORKSPACE_FRONTEND_URL_SCHEMA
 from tracim_backend.lib.utils.utils import get_frontend_ui_base_url
 from tracim_backend.lib.utils.utils import password_generator
-from tracim_backend.models import User
+from tracim_backend.models.auth import User
 from tracim_backend.models.auth import Group
 from tracim_backend.models.auth import Profile
 from tracim_backend.models.data import Content
