+++ backend/tracim_backend/models/revision_protection.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
+import typing
+
 from sqlalchemy.orm import Session
 from sqlalchemy import inspect
 from sqlalchemy.orm.unitofwork import UOWTransaction
@@ -7,7 +9,7 @@ from contextlib import contextmanager
 
 from tracim_backend.exceptions import ContentRevisionDeleteError
 from tracim_backend.exceptions import ContentRevisionUpdateError
-from tracim_backend.exceptions import SameValueError
+
 
 from tracim_backend.models.data import ContentRevisionRO
 from tracim_backend.models.data import Content
