+++ backend/tracim_backend/tests/commands/test_commands.py
@@ -17,9 +17,8 @@ from tracim_backend.exceptions import \
 from tracim_backend.exceptions import UserAlreadyExistError
 from tracim_backend.exceptions import UserDoesNotExist
 from tracim_backend.lib.core.user import UserApi
-from tracim_backend.models import get_engine
-from tracim_backend.models import get_session_factory
-from tracim_backend.models import get_tm_session
+from tracim_backend.models.setup_models import get_engine, get_session_factory, \
+    get_tm_session
 from tracim_backend.tests import CommandFunctionalTest
