+++ backend/tracim_backend/tests/functional/test_comments.py
@@ -1,14 +1,16 @@
 # -*- coding: utf-8 -*-
 import transaction
+
+from tracim_backend import error
 from tracim_backend.app_models.contents import content_type_list
-from tracim_backend.lib.core.content import ContentApi
-from tracim_backend.lib.core.workspace import WorkspaceApi
-from tracim_backend.models import get_tm_session
-from tracim_backend.models.revision_protection import new_revision
-from tracim_backend.tests import FunctionalTest
-from tracim_backend import error, models
 from tracim_backend.fixtures.content import Content as ContentFixtures
 from tracim_backend.fixtures.users_and_groups import Base as BaseFixture
+from tracim_backend.lib.core.content import ContentApi
+from tracim_backend.lib.core.workspace import WorkspaceApi
+from tracim_backend.models.auth import User
+from tracim_backend.models.revision_protection import new_revision
+from tracim_backend.models.setup_models import get_tm_session
+from tracim_backend.tests import FunctionalTest
 
 
 class TestCommentsEndpoint(FunctionalTest):
@@ -71,9 +73,9 @@ class TestCommentsEndpoint(FunctionalTest):
         Get alls comments of a content
         """
         dbsession = get_tm_session(self.session_factory, transaction.manager)
-        admin = dbsession.query(models.User) \
-            .filter(models.User.email == 'admin@admin.admin') \
-            .one() # type: models.User
+        admin = dbsession.query(User) \
+            .filter(User.email == 'admin@admin.admin') \
+            .one() # type: User
         workspace_api = WorkspaceApi(
             current_user=admin,
             session=dbsession,
@@ -140,9 +142,9 @@ class TestCommentsEndpoint(FunctionalTest):
         Get alls comments of a content
         """
         dbsession = get_tm_session(self.session_factory, transaction.manager)
-        admin = dbsession.query(models.User) \
-            .filter(models.User.email == 'admin@admin.admin') \
-            .one() # type: models.User
+        admin = dbsession.query(User) \
+            .filter(User.email == 'admin@admin.admin') \
+            .one() # type: User
         workspace_api = WorkspaceApi(
             current_user=admin,
             session=dbsession,
