+++ backend/tracim_backend/tests/models/test_content.py
@@ -10,9 +10,9 @@ import pytest
 # from tracim.lib.content import ContentApi
 from tracim_backend.exceptions import ContentRevisionUpdateError
 from tracim_backend.lib.core.content import ContentApi
-from tracim_backend.models import Content
+from tracim_backend.models.data import Content
 from tracim_backend.models.revision_protection import new_revision
-from tracim_backend.models import User
+from tracim_backend.models.auth import User
 from tracim_backend.models.data import ActionDescription
 from tracim_backend.models.data import ContentRevisionRO
 from tracim_backend.app_models.contents import content_type_list
